backend/requirements.txt
SQLAlchemy
alembic
psycopg2-binary 
python-dotenv 
fastapi
uvicorn.
firebase-admin
tzdata